Module BinaryEncoding.Buffer

type writer

Types of binary buffer

type reader

1. Writer

val initialize_writer : ?with_dict:bool -> unit -> writer

Initializes the buffer as a writer. If `with_dict` is set to true (by default `true`), strings will be registered in a dictionary (useful when there are multiple times the same string.

val finalize : writer -> Stdlib.Bytes.t

Returns the bytes representation of the buffer

val write : writer -> int64 -> int -> unit

`write buff v size` writes v as a 'size' bit value. Value must be positive.

val write_bool : writer -> bool -> unit

Writes a boolean on 1 bit

val write_uint8 : writer -> int -> unit

Write unsigned ints

val write_uint16 : writer -> int -> unit

val write_uint32 : writer -> int -> unit

val write_uint63 : writer -> int64 -> unit

val write_signed_int : writer -> int -> unit

Writes on the buffer an unsigned int prefixed by its size.

val write_z : writer -> Z.t -> unit

Writes an unbounded integer

val write_bytes : writer -> Stdlib.Bytes.t -> unit

Writes bytes

val write_string : writer -> string -> unit

Writes a string

val write_str_repr : writer -> string -> unit

Write a string representation

2. Reader

val initialize_reader : ?with_dict:bool -> Stdlib.Bytes.t -> reader

Initializes the buffer as a reader

val read : reader -> int -> int64

read buff n reads n bits on the buffer as an unsigned value

val read_bool : reader -> bool

Reads 1 bit, returns true if 1, 0 otherwise

val read_uint8 : reader -> int

Reads unsigned integers

val read_uint16 : reader -> int

val read_uint32 : reader -> int

val read_uint63 : reader -> int64

val read_signed_int : reader -> int

Reads an unsigned int prefixed by its size.

val read_z : reader -> Z.t

Reads an unbounded integer

val read_bytes : reader -> Stdlib.Bytes.t

Reads bytes

val read_string : reader -> string

Reads a string

val read_str_repr : reader -> string

Reads a string representation

val print_stats : Stdlib.Format.formatter -> writer -> unit

Statistics

val stats_add_i_custom : writer -> string -> int -> unit