Module Binary_encoding

module Buffer : module type of Buffer

The Buffer defines the methods for reading and writing in a buffer. The compression algorithm used by the library works by writing bits after bits instead of bytes.

module Lens : module type of Lens

Lenses are useful tools defining data-types encodings. You can define reader and writers that will be equivalent.

module Version : module type of Version