Module BinaryEncoding.Buffer

The Buffer defines the ùethods for reading and writing in a compressed buffer.

type reader

Writer

type writer

A writer is a Bytes buffer where is written the data. It also contains an optional dictionary for string storage. It also holds statistics

val initialize_writer : ?with_dict:bool -> ?init_size:int -> unit -> writer

Initializes a new buffer as a writer. If with_dict is set to true (by default true), strings will be registered in a dictionary (useful when the same string apprears at different places) and replaced by integers in the buffer. Variable init_size is the initial size of the writing bytes buffer (by default: 4096)

val finalize : writer -> Stdlib.Bytes.t

Returns the bytes representation of the buffer. If there is a dictionary, it is written at the beginning of the buffer.

val write : writer -> int64 -> int -> unit

write buff v size

Writes v on buff using size bits. Value must be positive. Fails with Invalid_argument if size is not in [0, 63] or if v is negative or greater than 2^size.

val write_bool : writer -> bool -> unit

Writes a boolean on 1 bit: 0 if false, 1 if true.

Write unsigned ints

val write_uint8 : writer -> int -> unit

Equivalent to write w v 8

val write_uint16 : writer -> int -> unit

Equivalent to write w v 16

val write_uint32 : writer -> int -> unit

Equivalent to write w v 32

val write_uint63 : writer -> int64 -> unit

Equivalent to write w v 63

val write_z : writer -> Z.t -> unit

Writes an unbounded integer.

val write_signed_int : writer -> int -> unit

Same as write_z w (Z.of_int i)

val write_bytes_known_length : len:int -> writer -> Stdlib.Bytes.t -> unit

write_bytes_known_size ~len w b Writes the first len bytes of b in argument.

val write_bytes : writer -> Stdlib.Bytes.t -> unit

Writes a bytes buffer of any size.

val write_str_repr : writer -> string -> unit

Write a string representation. If the buffer is associated to a dictionary, writes the its associated integer instead.

Reader

val initialize_reader : ?with_dict:bool -> Stdlib.Bytes.t -> reader

Initializes a buffer as a reader. If the writer used a dictionary, set with_dict to true (true by default).

val read : reader -> int -> int64

read buff n reads n bits on the buffer as an unsigned value

val read_bool : reader -> bool

Reads 1 bit, returns true if 1, false otherwise

Reads unsigned integers

val read_uint8 : reader -> int

Reads 8 bits

val read_uint16 : reader -> int

Reads 16 bits

val read_uint32 : reader -> int

Reads 32 bits

val read_uint63 : reader -> int64

Reads 63 bits

val read_signed_int : reader -> int

Reads an unsigned int prefixed by its size.

val read_z : reader -> Z.t

Reads an unbounded integer

val read_bytes_known_length : len:int -> reader -> bytes

write_bytes_constant_size ~len w b Reads len bytes.

val read_bytes : reader -> Stdlib.Bytes.t

Reads a integer representing the number of bytes to read, then reads and return these bytes.

val read_str_repr : reader -> string

Reads a string representation

Statistics

val print_stats : Stdlib.Format.formatter -> writer -> unit

Prints statistics of bits repartition of a buffer.